+++ app.py
@@ -1,11 +1,11 @@
 from flask import Flask, render_template, request, jsonify, redirect, flash, json, session
 import requests
 from forms import SearchRecipesForm
-from config import API_KEY
+from config import API_KEY, SECRET_KEY
 
 app = Flask(__name__)
 
-app.config['SECRET_KEY'] = 'MY_SECRET_KEY'
+app.config['SECRET_KEY'] = 'SECRET_KEY'
 API_BASE_URL = "https://api.spoonacular.com"
 
 @app.route('/', methods=['GET', 'POST'])
@@ -19,6 +19,7 @@ def show_recipes():
     response = requests.get(f"{API_BASE_URL}/food/ingredients/search", params={"apiKey": API_KEY, "query": ingredient})
 
     data = response.json()
+
     results = data['results']
 
     return render_template('search-results.html', data=data)
